requirements.txt
streamlit
chromadb
langchain
langchain-ollama
langchain-community
langchain-text-splitters
pypdf2
requests
ollama
python-docx
pandas
